Split Text Generation request into two parts: 1) Request generation 2) Request result + Other stuff

+++ api/src/main/kotlin/com/momiji/api/neural/text/model/HistoryRequest.kt
@@ -14,7 +14,7 @@ data class HistoryRequest(
     @JsonProperty("num_beams")
     val numBeams: Int? = null,
     @JsonProperty("repetition_penalty")
-    val repetitionPenalty: Int? = null,
+    val repetitionPenalty: Float? = null,
     @JsonProperty("early_stopping")
     val earlyStopping: Boolean? = null,
     @JsonProperty("seed")
